requirements.txt
streamlit
openai
numpy
python-dotenv
pypdf